added layout for blog
